new try

+++ heap.c
@@ -39,9 +39,9 @@ void heap_push(Heap* pq, void* data, int priority){
   int y =pq->size;
   while(1){
     if (pq->heapArray[((y-1)/2)]->priority < new->priority){
-      pq->heapArray[y]= pq->heapArray[int((y-1)/2)];
+      pq->heapArray[y]= pq->heapArray[((y-1)/2)];
       pq->heapArray[((y-1)/2)] = new;
-      y = int((y-1)/2);
+      y = ((y-1)/2);
       if (y<=0) break;
     }
     else break;
